[fix]Rangeの変更をBaseで行うように変更
