+++ Assets/Master/Resources/ScriptableObject/DefenseEquipmentData/DataBase/ShootTurretDataBase.asset
@@ -16,4 +16,4 @@ MonoBehaviour:
   Explanation: 
   DropChance: 0
   DataLevelList:
-  - {fileID: 11400000, guid: 62325683173c228468ea3cbe04a5a4d5, type: 2}
+  - {fileID: 11400000, guid: 9dac64c79e2548b43a20878ffd3629bf, type: 2}
