--- Assets/Master/Resources/ScriptableObject/DefenseEquipmentData/Datas/ShootTurretDataLevel1.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 72693d83c78b97c468b889e7e5db6cec, type: 3}
-  m_Name: ShootTurretDataLevel1
-  m_EditorClassIdentifier: 
-  Attack: 2
-  Rate: 2
-  Critical: 30
-  Range: 6
