--- Assets/Master/Resources/ScriptableObject/DefenseEquipmentData/Datas/ShootTurretDataLevel1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 62325683173c228468ea3cbe04a5a4d5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
