+++ Assets/Master/Script/InGame/Tallet/DefenseEquipmentManager_S/ShortRangeShootTurretManager.cs.meta
