+++ Assets/Master/Script/InGame/Tallet/DefenseEquipmentManager_S/SummonEntityMasterManagar.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+namespace DCFrameWork
+{
+    public class SummonEntityMasterManagar : MonoBehaviour
+    {
+        // Start is called once before the first execution of Update after the MonoBehaviour is created
+        void Start()
+        {
+        
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+    }
+}
